+++ src/main/java/com/github/unchama/seichiassist/Sql.java
@@ -4,10 +4,10 @@ import com.github.unchama.seichiassist.data.GachaData;
 import com.github.unchama.seichiassist.data.MineStackGachaData;
 import com.github.unchama.seichiassist.data.PlayerData;
 import com.github.unchama.seichiassist.data.RankData;
-import com.github.unchama.seichiassist.database.init.DonateDataTableQueryGenerator;
-import com.github.unchama.seichiassist.database.init.GachaDataTableQueryGenerator;
-import com.github.unchama.seichiassist.database.init.MineStackGachaDataTableQueryGenerator;
-import com.github.unchama.seichiassist.database.init.PlayerDataTableQueryGenerator;
+import com.github.unchama.seichiassist.database.ddl.DonateDataTableQueryGenerator;
+import com.github.unchama.seichiassist.database.ddl.GachaDataTableQueryGenerator;
+import com.github.unchama.seichiassist.database.ddl.MineStackGachaDataTableQueryGenerator;
+import com.github.unchama.seichiassist.database.ddl.PlayerDataTableQueryGenerator;
 import com.github.unchama.seichiassist.task.CheckAlreadyExistPlayerDataTaskRunnable;
 import com.github.unchama.seichiassist.task.CoolDownTaskRunnable;
 import com.github.unchama.seichiassist.task.PlayerDataSaveTaskRunnable;
